+++ src/routes.js
@@ -7,8 +7,9 @@
 import { game } from 'constants/localize/es';
 
 export type RouteDescription = {|
-  game: string,
+  difficulty: 'easy' | 'normal' | 'hard',
   exact: boolean,
+  game: string,
   title: string,
   children: Array<{|
     title: string,
@@ -18,33 +19,47 @@ export type RouteDescription = {|
 const routes: Array<RouteDescription> = [{
   exact: true,
   game: 'maze',
+  difficulty: 'easy',
   title: game.difficulty.easy,
   children: [{
     title: `${game.levels[0]} ${game.exercise}`,
-    path: 'easy/001',
+    path: '001',
   }, {
     title: `${game.levels[1]} ${game.exercise}`,
-    path: 'easy/002',
+    path: '002',
   }, {
     title: `${game.levels[2]} ${game.exercise}`,
-    path: 'easy/003',
+    path: '003',
   }, {
     title: `${game.levels[3]} ${game.exercise}`,
-    path: 'easy/004',
+    path: '004',
   }],
 }, {
   exact: true,
   game: 'maze',
+  difficulty: 'normal',
   title: game.difficulty.normal,
   children: [{
     title: `${game.levels[0]} ${game.exercise}`,
-    path: 'normal/001',
+    path: '001',
   }, {
     title: `${game.levels[1]} ${game.exercise}`,
-    path: 'normal/002',
+    path: '002',
   }, {
     title: `${game.levels[2]} ${game.exercise}`,
-    path: 'normal/003',
+    path: '003',
+  }],
+}, {
+  exact: true,
+  game: 'maze',
+  difficulty: 'hard',
+  title: game.difficulty.hard,
+  children: [{
+    title: `${game.levels[0]} ${game.exercise}`,
+    path: '001',
+  }, {
+    title: `${game.levels[1]} ${game.exercise}`,
+    path: '002',
   }],
 }];